Basic container starting plumbing
Can start a container from an image.  If the image doesn't exists,
we'll pull it down.  Private registries use .dockercfg auth that
is setup via docker login.  After starting the new container, any
versions of that container still running that are older than the
cutoff period are shut down.

We'd leave the old container running for a bit to:

  1) Allow the new container address to propogate out to the other hosts
  2) Allow running work in that container to finish
  3) Make rollbacks easier