Change etcd tree to support atomic updates
Using an entry per key would cause watches to get notified of each key
change which complicates watcher implementations.  For example, the
proxy update piece would get a change for a new port, then one for an ip
and would need to buffer the updates.  Using JSON as the value would
fixes that.

Along with this, we changed the keys in etcd to be all lowercase.  The
json value for the config entries will have uppercase keys to they
more closely resemble environment variables.