Re-work how old containers are stopped
Previously, old containers would only be stopped if a new container
was started and the old containers happened to be older than the
cutoff threshold.  This left old containers around inadvertently.

Containers are now looked at on a schedule to determine if they
should be stopped.  The stop threshold has also been lowered to
to stop old containers more aggressively.