Start containers that should be running
Containers would only be started when commander was started.  If
a containers was stopped via the docker CLI, it would not get
restarted until commander was restarted.  Ideally, we would use
the docker events API to detect these changes but for now this will
just start any containers on the existing scheduled timer.