Fix ServiceConfig.ID not always incrementing
Since ServiceConfig is composed of several VMaps, it's ID is
the max of the IDs.  If the environment was updated frequently,
it's version might be, say 10 which would make the ServiceConfig.ID,
10 as well.  If a deploy occurred, the version VMap might only be
at 2 which would be less than the current version so the ID would
not get incremented.  This caused some changes to not get picked
up.

The change forces all updates to use the next version ID for the
ServiceConfig.