Replace container naming code with function
How galaxy containers were named was spread throughout a few places.
This also fixes a few bugs when app:shell runs.  One is that
commander would get confused and try to start a container that
is already running when a shell is opened.  It would fail. Another
was that discovery would register/unregister/report status on these
temporary containers when they should be ignored.

Using container names is kind of fragile and hacky but I'm not sure
of a better way to identify galaxy managed containers from other
ones that might be running.