Fix old containers not getting removed on re-deploys
When deploying by updating the "latest" tag in the registry,
old containers would get left running because they would
become unassociated with the registry tag and commander would
ignore them.  This changes that to rely on the container name
and image ID which should be more reliable to determining if
a container should be running or not.