Always pull image when config has changed
This should fix auto-deploys and allow us to use docker caching
again when deploying images.  Auto-deploys were actually pulling
the image on the build node and not the runtime nodes because
galaxy was running in local mode and configured w/ the envs
redis instance.  When the deploy completed, the runtime nodes
just recreated the container w/ the same image.

This change forces commander to always pull down the latest image when
it's notified of a config change.