Pull images in parallel at start
Commander would pull each image serially which is pretty slow for
large images.  This attempts to load them in parallel and let the
docker daemon download them as quickly as it can.

It's still not quite optimal in that some containers could be started
while others are still being downloaded.