Replace ADDR_ env vars w/ generic HOST_IP
We were creating env vars for container such as ADDR_APIARY_5000
which would point to the the shuttle port on the host which would
forward to an apiary instance in the env.  It was kind of clunky
and didn't work for some cases so just set a generic HOST_IP
on containers and let them determine what ports they want to connect
to on the host.

This will make it possible for container to hit dogstatsd as well
as shuttle ports w/ one config which is simpler to understand and
explain.