Make service configs env specific
ServiceConfigs (app version, env) were scoped to an env and pool
per service.  If we have multiple pools in an environment and a
service is deployed across both (web + worker), we'd have to duplicate
deployment and config variables which will likely cause a lot of
issues.

Remove the pool scoping so they are consistent within an environment.

This has some rename logic that will update the current keys to
avoid some manual work on the redis host.