Set STATSD_ADDR env var on started containers
The HOST_IP can be used by containers but the original
idea of setting a STATSD_ADDR=$HOST_IP on the container
does not work because docker escape the $ char in the
env var value.

It probably be better to be able to pass arbitrary flags
to commander w/ something like -e STATSD_ADDR=xyz but the
current golang flags package doesn't allow setting the same
flag param multiple times.

This patch just sets a STATSD_ADDR in the same what HOST_IP is
set as a workaround until a better solution can be implemented.