Set  PUBLIC_HOSTNAME env var on containers
This can be used by cesta to set the correct upload endpoint.  For
now, it just returns the EC2 public DNS entry but it will eventually
have a evermore.photos CNAME value.  SSL will not work yet but this
should make it possible to upload over HTTP directly to the host.