Initial quasar http proxy server
This is a frontend reverse proxy and load balancer that uses galaxy service
registrations to setup backend routes.  The current implementation using
nginx has a few issues when new hosts boot up.  The first is that
the ELB checks a status served by nginx to determine if the host
is ready for requests.  Unfortunately, nginx is usually up and ready
but the docker containers running services are not.  This results
in a period of random 503 for each service depending on how many hosts
are enabled in the ELB.

quasar will address this by routing requests to other hosts if the host
it runs on does not have the service available.  In the future, we can set
priority to route to the local service and fallback to a remote host
if needed.