New method for proxy dialing
We now give the reverse proxy a list of addresses it can try for the
backends. These are sent through our service Dialer so they can be
looked up against the real backends, and hooked into the stats. Dialing
can return a DialError in the case that the connection could not be
established, so that the proxy knows it's safe to attempt another
backend.

Also fixes request logging.