Add some host registration commands
When commander starts on a host, the host is registered
with it's IP.  Eventually, other meta-data will be registered
as well so we can get better insight into the cluster as
well as more fine-grained control of how containers are
run on each host.

Initially this will just be used to implement some basic
instance balancing across hosts so we can have galaxy
run 10 instances of an app across 5 hosts (2 on each host). This
will also be useful for scaling containers up/down across hosts.