Don't depend on explict container names
We required that containers be name <app>_<version> to determine
if we managed them and whether they were running the correct
version.  This limits how many instances of a container we can
run on a given host.

We now name containers anything and instead rely on container
metadata set through env variables.