Add a hostIP to runtime
It's was overloaded w/ shuttleAddr so separate the two.  Also
adding support for env vars of the form FOO=$HOST_IP:1234 which
will be replaced at runtime w/ the value of HOST_IP.  This will
allow removing the statsdAddr and shuttleAddr options which
can be set w/ regular env vars instead.