commander: use rolling restart during deploy
When multiple processe are configured a deploy can create a lot
of containers during deploy.  For example, with 10 running, it would
start 10 new ones and then kill the 10 old ones.  This now starts
one and stops and old one to keep the number of containers more
per host more stable.