commander: fix shuttle services being repeated add/removed
If more than one pool exists, commanders in each pool would add
the shuttle routes for the services in the other pools.  It would
then see that none of those services were running in it's pool
and subsequently remove it causing them to be repeatedly added
and removed.

Changed the remove logic to remove the service from shuttle if
it no longer exists in the registry.  Backends are removed as
before.